LayoutTests/webgpu/whlsl-buffer-length-expected.txt

PASS resultUint32Array[0] is 1337
PASS successfullyParsed is true

TEST COMPLETE
